refactor(frontend): elevate theme context to app-level providers
